cat-cafe-backup/requirements.txt
# Cat Café Multi-Agent System - Dependencies

# Core dependencies
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyYAML>=6.0

# Storage dependencies
redis>=4.0.0
SQLAlchemy>=2.0.0
asyncpg>=0.28.0  # PostgreSQL async driver

# WebSocket support
websockets>=11.0

# Utility dependencies
python-dotenv>=1.0.0

# Development dependencies (optional)
black>=23.0.0
isort>=5.0.0
pylint>=3.0.0
pytest>=7.0.0
mypy>=1.0.0
